src/test/javascript/spec/app/entities/tipo-servicio/tipo-servicio.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { TipoServicioService } from 'app/entities/tipo-servicio/tipo-servicio.service';
import { ITipoServicio, TipoServicio, TipoRecurso } from 'app/shared/model/tipo-servicio.model';

describe('Service Tests', () => {
    describe('TipoServicio Service', () => {
        let injector: TestBed;
        let service: TipoServicioService;
        let httpMock: HttpTestingController;
        let elemDefault: ITipoServicio;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(TipoServicioService);
            httpMock = injector.get(HttpTestingController);

            elemDefault = new TipoServicio(0, 'AAAAAAA', 'AAAAAAA', false, TipoRecurso.BAHIA);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign({}, elemDefault);
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a TipoServicio', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .create(new TipoServicio(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a TipoServicio', async () => {
                const returnedFromService = Object.assign(
                    {
                        nombre: 'BBBBBB',
                        codigo: 'BBBBBB',
                        interno: true,
                        tipoRecurso: 'BBBBBB'
                    },
                    elemDefault
                );

                const expected = Object.assign({}, returnedFromService);
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of TipoServicio', async () => {
                const returnedFromService = Object.assign(
                    {
                        nombre: 'BBBBBB',
                        codigo: 'BBBBBB',
                        interno: true,
                        tipoRecurso: 'BBBBBB'
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a TipoServicio', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
